installerFiles/Linux/Agent/mcp/requirements.txt
pyyaml>=6.0
nats-py>=2.11.0
pytest>=7.0
pytest-asyncio>=0.21.0
